requirements.txt
Django==2.1.3
channels==2.1.5
channels_redis==2.3.1
uvicorn==0.4.1
wsproto==0.12.0

websocket-client